Incorporate chat into main.dart
